feat(llama-api-server): new `utils` mod
