+++ api-server/llama-api-server/src/utils.rs
@@ -1,0 +1,33 @@
+use url::Url;
+
+pub(crate) fn print_log_begin_separator(
+    title: impl AsRef<str>,
+    ch: Option<&str>,
+    len: Option<usize>,
+) -> usize {
+    let title = format!(" [LOG: {}] ", title.as_ref());
+
+    let total_len: usize = len.unwrap_or(100);
+    let separator_len: usize = (total_len - title.len()) / 2;
+
+    let ch = ch.unwrap_or("-");
+    let mut separator = "\n\n".to_string();
+    separator.push_str(ch.repeat(separator_len).as_str());
+    separator.push_str(&title);
+    separator.push_str(ch.repeat(separator_len).as_str());
+    separator.push('\n');
+    println!("{}", separator);
+    total_len
+}
+
+pub(crate) fn print_log_end_separator(ch: Option<&str>, len: Option<usize>) {
+    let ch = ch.unwrap_or("-");
+    let mut separator = "\n\n".to_string();
+    separator.push_str(ch.repeat(len.unwrap_or(100)).as_str());
+    separator.push_str("\n\n");
+    println!("{}", separator);
+}
+
+pub(crate) fn is_valid_url(url: &str) -> bool {
+    Url::parse(url).is_ok()
+}
